fixup! Added mnist normalization and complex numbers pooling

+++ src/geoconv_examples/mnist/dataset.py
@@ -21,10 +21,10 @@ def dataset(mnist_atlas, set_type, n_radial, n_angular, batch_size):
     # Return image, barycentric coordinates and label
     def transform(image, label):
         # Normalize image
-        image = image / 255.
+        image = tf.cast(image, tf.float32) / 255.
 
         # Lift 1-d features into imaginary plane by mapping 'grey_scale_value -> grey_scale_value + 0i'
-        image = tf.reshape(tf.cast(image, tf.float32), (784, 1))
+        image = tf.reshape(image, (784, 1))
         image = tf.concat([image, imaginary_values], axis=-1)
         return (image, bc, rotations, rotation_orders), label
     mnist = mnist.map(transform)
